Set up structure of mkdocs parser class
