+++ mkdocs/src/main.py
@@ -8,7 +8,9 @@ from visivo.models.defaults import Defaults
 from visivo.models.item import Item
 from visivo.models.row import Row
 from visivo.models.alert import Alert, SlackAlert, EmailAlert
-from visivo.parsers.defs_generator import generate_defs
+from visivo.parsers.mkdocs.defs_generator import generate_defs
+#from .markdown_generators import general_pydantic as gp
+from visivo.parsers.mkdocs.markdown import  from_pydantic_model, is_pydantic_model
 from visivo.models.trace_props import (
     TraceProps,
     Mesh3d,
@@ -63,31 +65,6 @@ import yaml
 from typing import List, Union, Dict, Any, Type
 
 from mkdocs_click._extension import replace_command_docs
-
-def _return_model(model_name: str) -> BaseModel:
-    match model_name:
-        case 'trace':
-            return Trace
-        case 'target':
-            return Target
-        case 'test':
-            return Test
-        case 'chart':
-            return Chart
-        case 'dashboard':
-            return Dashboard
-        case 'default':
-            return Defaults
-        case 'item':
-            return Item
-        case 'row':
-            return Row
-        case 'alert':
-            return Alert
-        case 'slack-alert':
-            return SlackAlert
-        case 'email-alert':
-            return EmailAlert
 
 def _return_trace_prop_model(model_name: str) -> TraceProps:
     match model_name:
@@ -183,10 +160,6 @@ def _return_trace_prop_model(model_name: str) -> TraceProps:
             return Scattergl
         case 'splom':
             return Splom
-        
-def is_pydantic_model(cls) -> bool:
-    """Check if a given class is a Pydantic model."""
-    return isinstance(cls, type) and issubclass(cls, BaseModel)
 
 def extract_model_info(model: Type[BaseModel]) -> Dict[str, Any]:
     output = {}
@@ -220,25 +193,6 @@ def generate_yaml_from_model(model: Type[BaseModel]) -> str:
 
 
 def define_env(env):
-    def pydantic_model_to_md_table(model_name: str):
-        """Generates markdown tables for pydantics models"""
-        model = _return_model(model_name)
-        fields = model.__fields__
-        model_md = '' if model.__doc__  == None else dedent(model.__doc__)        
-        md_table = "| Field | Type | Default | Description |\n|-------|------|---------|-------------|\n"
-
-        def clean_field_type(field):
-            field_annotation = str(field.annotation)
-            return field_annotation.replace('typing.', '')
-
-        for field_name, field in fields.items():
-            field_type = clean_field_type(field=field)
-            field_description = field.description
-            field_default = '-' if field.default == None else field.default
-            md_table += f"| {field_name} | {field_type} | {field_default} | {field_description} |\n"
-
-        return model_md + '\n' + md_table
-    
     env.macro(pydantic_model_to_md_table, "render_pydantic_model")
 
     def render_click_docs( has_attr_list= False, options= {}):
