+++ tests/mkdocs/test_main.py
@@ -1,4 +1,4 @@
-from mkdocs.src.macros.main import is_pydantic_model
+from mkdocs.src.main import is_pydantic_model
 from visivo.models.trace import Trace
