+++ tests/parsers/test_mkdocs/test_markdown.py
@@ -1,0 +1,5 @@
+from visivo.parsers.mkdocs import markdown
+
+
+def test_markdown_generation():
+    markdown.from_pydantic_model()
