+++ visivo/parsers/mkdocs_utils/defs_generator.py
@@ -1,4 +1,4 @@
-from .schema_generator import generate_schema
+from ..schema_generator import generate_schema
 import json
 
 
@@ -6,5 +6,5 @@ def generate_defs() -> dict:
     schema = json.loads(generate_schema())
     defs = schema.get('$defs', {})
     if not defs:
-        raise Exception(f"Schema is missing '$defs' key. ")
+        raise Exception(f"Schema is missing '$defs' key. Maybe pydantic made some breaking updates?")
     return defs
